Move server management tests to their own file. Add gradle test.
