+++ ycmd/tests/java/subcommands_test.py
@@ -22,15 +22,12 @@ from __future__ import division
 # Not installing aliases from python-future; it's unreliable and slow.
 from builtins import *  # noqa
 
-from hamcrest import (
-  assert_that,
-  contains,
-  contains_inanyorder,
-  empty,
-  has_entries,
-  has_entry,
-  instance_of,
-)
+from hamcrest import ( assert_that,
+                       contains,
+                       contains_inanyorder,
+                       empty,
+                       has_entries,
+                       instance_of )
 from nose.tools import eq_
 from pprint import pformat
 import requests
@@ -838,167 +835,3 @@ def Subcommands_FixIt_NoDiagnostics_test():
 
   yield ( RunFixItTest, "no FixIts means you gotta code it yo' self",
           filepath, 1, 1, has_entries( { 'fixits': empty() } ) )
-
-
-@IsolatedYcmdInDirectory( PathToTestFile( 'simple_eclipse_project' ) )
-def Subcommands_RestartServer_test( app ):
-  WaitUntilCompleterServerReady( app )
-
-  eclipse_project = PathToTestFile( 'simple_eclipse_project' )
-  maven_project = PathToTestFile( 'simple_maven_project' )
-
-  # Run the debug info to check that we have the correct project dir
-  request_data = BuildRequest( filetype = 'java' )
-  assert_that(
-    app.post_json( '/debug_info', request_data ).json,
-    has_entry( 'completer', has_entries( {
-      'name': 'Java',
-      'servers': contains( has_entries( {
-        'name': 'Java Language Server',
-        'is_running': instance_of( bool ),
-        'executable': instance_of( str ),
-        'pid': instance_of( int ),
-        'logfiles': contains( instance_of( str ),
-                              instance_of( str ) ),
-        'extras': contains(
-          has_entries( { 'key': 'Java Path',
-                         'value': instance_of( str ) } ),
-          has_entries( { 'key': 'Launcher Config.',
-                         'value': instance_of( str ) } ),
-          has_entries( { 'key': 'Project Directory',
-                         'value': eclipse_project } ),
-          has_entries( { 'key': 'Workspace Path',
-                         'value': instance_of( str ) } )
-        )
-      } ) )
-    } ) )
-  )
-
-  # Restart the server with a different client working directory
-  filepath = PathToTestFile( 'simple_maven_project',
-                             'src',
-                             'main',
-                             'java',
-                             'com',
-                             'test',
-                             'TestFactory.java' )
-
-  app.post_json(
-    '/run_completer_command',
-    BuildRequest(
-      filepath = filepath,
-      filetype = 'java',
-      working_dir = maven_project,
-      command_arguments = [ 'RestartServer' ],
-    ),
-  )
-
-  WaitUntilCompleterServerReady( app )
-
-  app.post_json(
-    '/event_notification',
-    BuildRequest(
-      filepath = filepath,
-      filetype = 'java',
-      working_dir = maven_project,
-      event_name = 'FileReadyToParse',
-    )
-  )
-
-  # Run the debug info to check that we have the correct project dir
-  request_data = BuildRequest( filetype = 'java' )
-  assert_that(
-    app.post_json( '/debug_info', request_data ).json,
-    has_entry( 'completer', has_entries( {
-      'name': 'Java',
-      'servers': contains( has_entries( {
-        'name': 'Java Language Server',
-        'is_running': instance_of( bool ),
-        'executable': instance_of( str ),
-        'pid': instance_of( int ),
-        'logfiles': contains( instance_of( str ),
-                              instance_of( str ) ),
-        'extras': contains(
-          has_entries( { 'key': 'Java Path',
-                         'value': instance_of( str ) } ),
-          has_entries( { 'key': 'Launcher Config.',
-                         'value': instance_of( str ) } ),
-          has_entries( { 'key': 'Project Directory',
-                         'value': maven_project } ),
-          has_entries( { 'key': 'Workspace Path',
-                         'value': instance_of( str ) } )
-        )
-      } ) )
-    } ) )
-  )
-
-
-@IsolatedYcmdInDirectory( PathToTestFile( 'simple_eclipse_project', 'src' ) )
-def Subcommands_ProjectDetection_EclipseParent( app ):
-  WaitUntilCompleterServerReady( app )
-
-  project = PathToTestFile( 'simple_eclipse_project' )
-
-  # Run the debug info to check that we have the correct project dir
-  request_data = BuildRequest( filetype = 'java' )
-  assert_that(
-    app.post_json( '/debug_info', request_data ).json,
-    has_entry( 'completer', has_entries( {
-      'name': 'Java',
-      'servers': contains( has_entries( {
-        'name': 'Java Language Server',
-        'is_running': instance_of( bool ),
-        'executable': instance_of( str ),
-        'pid': instance_of( int ),
-        'logfiles': contains( instance_of( str ),
-                              instance_of( str ) ),
-        'extras': contains(
-          has_entries( { 'key': 'Java Path',
-                         'value': instance_of( str ) } ),
-          has_entries( { 'key': 'Launcher Config.',
-                         'value': instance_of( str ) } ),
-          has_entries( { 'key': 'Project Directory',
-                         'value': project } ),
-          has_entries( { 'key': 'Workspace Path',
-                         'value': instance_of( str ) } )
-        )
-      } ) )
-    } ) )
-  )
-
-
-@IsolatedYcmdInDirectory( PathToTestFile( 'simple_maven_project',
-                                          'src',
-                                          'java',
-                                          'test' ) )
-def Subcommands_ProjectDetection_MavenParent( app ):
-  WaitUntilCompleterServerReady( app )
-
-  project = PathToTestFile( 'simple_maven_project' )
-
-  # Run the debug info to check that we have the correct project dir
-  request_data = BuildRequest( filetype = 'java' )
-  assert_that(
-    app.post_json( '/debug_info', request_data ).json,
-    has_entry( 'completer', has_entries( {
-      'name': 'Java',
-      'servers': contains( has_entries( {
-        'name': 'Java Language Server',
-        'is_running': instance_of( bool ),
-        'executable': instance_of( str ),
-        'pid': instance_of( int ),
-        'logfiles': contains( instance_of( str ),
-                              instance_of( str ) ),
-        'extras': contains(
-          has_entries( { 'key': 'Java Path',
-                         'value': instance_of( str ) } ),
-          has_entries( { 'key': 'Launcher Config.',
-                         'value': instance_of( str ) } ),
-          has_entries( { 'key': 'Project Directory',
-                         'value': project } ),
-          has_entries( { 'key': 'Workspace Path',
-                         'value': instance_of( str ) } )
-        )
-      } ) )
-    } ) )
-  )
